Reorganize  interfaces, fix mocking problem
