+++ client/src/preview/route/digitaltwins/execute/pipelineChecks.ts
@@ -15,7 +15,7 @@ import {
   isSuccessStatus,
   isFailureStatus,
 } from 'model/backend/gitlab/execution/statusChecking';
-import { ProjectId } from 'model/backend/interfaces/utilityInterfaces';
+import { ProjectId } from 'model/backend/interfaces/backendInterfaces';
 
 interface PipelineStatusParams {
   setButtonText: Dispatch<SetStateAction<string>>;
