+++ client/src/preview/util/DTAssets.ts
@@ -1,12 +1,13 @@
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable no-await-in-loop */
 
+import { BackendInterface } from 'model/backend/interfaces/backendInterfaces';
 import {
-  BackendInterface,
   DTAssetsInterface,
   FileHandlerInterface,
-} from 'model/backend/interfaces/utilityInterfaces';
-import { FileState, FileType } from 'model/backend/interfaces/sharedInterfaces';
+  FileState,
+  FileType,
+} from 'model/backend/interfaces/sharedInterfaces';
 import FileHandler from './fileHandler';
 
 export function getFilePath(
