+++ client/src/preview/util/digitalTwinUtils.ts
@@ -2,15 +2,15 @@
 
 import {
   BackendAPI,
-  LibraryConfigFile,
   ProjectId,
   RepositoryTreeItem,
-} from 'model/backend/interfaces/utilityInterfaces';
+} from 'model/backend/interfaces/backendInterfaces';
+import { LibraryConfigFile } from 'model/backend/interfaces/sharedInterfaces';
 import { Asset } from 'preview/components/asset/Asset';
 import { AssetTypes } from 'model/backend/gitlab/digitalTwinConfig/constants';
 import { getDTDirectory } from 'model/backend/gitlab/digitalTwinConfig/settingsUtility';
 import GitlabInstance from 'model/backend/gitlab/instance';
-import { ExecutionStatus } from 'model/backend/gitlab/types/executionHistory';
+import { ExecutionStatus } from 'model/backend/interfaces/execution';
 import DigitalTwin from './digitalTwin';
 
 export function isValidInstance(digitalTwin: DigitalTwin): boolean {
