+++ client/src/preview/util/libraryManager.ts
@@ -1,11 +1,12 @@
 /* eslint-disable no-restricted-syntax */
 /* eslint-disable no-await-in-loop */
 
+import { BackendInterface } from 'model/backend/interfaces/backendInterfaces';
 import {
-  BackendInterface,
   LibraryManagerInterface,
-} from 'model/backend/interfaces/utilityInterfaces';
-import { FileState, FileType } from 'model/backend/interfaces/sharedInterfaces';
+  FileState,
+  FileType,
+} from 'model/backend/interfaces/sharedInterfaces';
 import FileHandler from './fileHandler';
 
 export function getFilePath(
