+++ client/test/preview/__mocks__/constants.mock.ts
@@ -3,15 +3,12 @@
  * This centralizes all the mocks for the constants module in one place
  * so they can be imported consistently across tests
  */
-
-import { getBranchName } from 'model/backend/gitlab/digitalTwinConfig/settingsUtility';
-
 const DEFAULT_MOCK_VALUES = {
   GROUP_NAME: 'DTaaS',
   DT_DIRECTORY: 'digital_twins',
   COMMON_LIBRARY_PROJECT_NAME: 'common',
   RUNNER_TAG: 'linux',
-  BRANCH_NAME: getBranchName(),
+  BRANCH_NAME: 'master',
 };
 
 /**
@@ -19,27 +16,60 @@ const DEFAULT_MOCK_VALUES = {
  * @param overrides - Override specific values from the defaults
  * @returns The mock object that can be used with jest.mock
  */
+const createSettingsUtilityMock = (overrides = {}) => {
+  const mockValues = { ...DEFAULT_MOCK_VALUES, ...overrides };
+
+  return {
+    getGroupName: jest.fn().mockReturnValue(mockValues.GROUP_NAME),
+    getDTDirectory: jest.fn().mockReturnValue(mockValues.DT_DIRECTORY),
+    getCommonLibraryProjectName: jest
+      .fn()
+      .mockReturnValue(mockValues.COMMON_LIBRARY_PROJECT_NAME),
+    getRunnerTag: jest.fn().mockReturnValue(mockValues.RUNNER_TAG),
+    getBranchName: jest.fn().mockReturnValue(mockValues.BRANCH_NAME),
+  };
+};
+
+/**
+ * Create a mock for the constants module
+ */
 const createConstantsMock = (overrides = {}) => {
   const mockValues = { ...DEFAULT_MOCK_VALUES, ...overrides };
 
   return {
-    // Non-hook getter functions
-    getGroupName: jest.fn().mockReturnValue(mockValues.GROUP_NAME),
-    getDTDirectory: jest.fn().mockReturnValue(mockValues.DT_DIRECTORY),
-    getCommonLibraryProjectId: jest
-      .fn()
-      .mockReturnValue(mockValues.COMMON_LIBRARY_PROJECT_NAME),
-    getRunnerTag: jest.fn().mockReturnValue(mockValues.RUNNER_TAG),
-
-    // Hook versions if needed
-    useGroupName: jest.fn().mockReturnValue(mockValues.GROUP_NAME),
-    useDTDirectory: jest.fn().mockReturnValue(mockValues.DT_DIRECTORY),
-    useCommonLibraryProjectId: jest
-      .fn()
-      .mockReturnValue(mockValues.COMMON_LIBRARY_PROJECT_NAME),
-    useRunnerTag: jest.fn().mockReturnValue(mockValues.RUNNER_TAG),
-    useBranchName: jest.fn().mockReturnValue(mockValues.BRANCH_NAME),
+    GROUP_NAME: mockValues.GROUP_NAME,
+    DT_DIRECTORY: mockValues.DT_DIRECTORY,
+    COMMON_LIBRARY_PROJECT_NAME: mockValues.COMMON_LIBRARY_PROJECT_NAME,
+    RUNNER_TAG: mockValues.RUNNER_TAG,
+    BRANCH_NAME: mockValues.BRANCH_NAME,
+    MAX_EXECUTION_TIME: 10 * 60 * 1000,
+    PIPELINE_POLL_INTERVAL: 5 * 1000,
+    AssetTypes: {
+      Functions: 'functions',
+      Models: 'models',
+      Tools: 'tools',
+      Data: 'data',
+      'Digital Twins': 'digital_twins',
+      'Digital Twin': 'digital_twin',
+    },
+    defaultFiles: [
+      { name: 'description.md', type: 'DESCRIPTION' },
+      { name: 'README.md', type: 'DESCRIPTION' },
+      { name: '.gitlab-ci.yml', type: 'CONFIGURATION' },
+    ],
   };
 };
 
-export default createConstantsMock;
+export const settingsUtilityMock = createSettingsUtilityMock();
+export const constantsMock = createConstantsMock();
+
+export { createSettingsUtilityMock, createConstantsMock };
+
+jest.mock(
+  'model/backend/gitlab/digitalTwinConfig/settingsUtility',
+  () => settingsUtilityMock,
+);
+jest.mock(
+  'model/backend/gitlab/digitalTwinConfig/constants',
+  () => constantsMock,
+);
