+++ client/test/preview/unit/jest.setup.ts
@@ -1,11 +1,10 @@
 import '@testing-library/jest-dom';
 import 'test/preview/__mocks__/global_mocks';
-import 'test/preview/__mocks__/constants.mock';
 import 'test/__mocks__/global_mocks';
 import 'test/__mocks__/unit/page_mocks';
 import 'test/__mocks__/unit/component_mocks';
 import 'test/__mocks__/unit/module_mocks';
 
 beforeEach(() => {
-  jest.resetAllMocks();
+  jest.clearAllMocks();
 });
